refactor: moving Token to a different package
Moving TokenRequest and TokenResponse to "token" package
